Fix use of String.format in BashCommands
Should not really be used with no arguments, and should *definitely* never be used when the format string may contain '%'
characters, which is always a possibility with URLs. This was causing `JBoss6SshDriver.install` to fail, since there are
URL-escaped characters in the download URL, when it eventually called `BashCommands.alternatives` which ended up doing a
superfluous `format` on the list of commands with no arguments.