Expose override of SSH host and port; use w/ knife
SshMachineLocation allows the hostname and port number for SSH access to be overridden, while still having the main MachineLocation address unchanged. This is useful for cases where the "address" is a private IP that is not reachable by the Brooklyn management console, but a firewall/port-forwarding configuration allows it to be SSH-accessible on an alternative hostname and port number. Adds a public method that retrieves the SSH host and port.

Alters the "knife converge" task to extract this alternative host-and-port and use it in the bootstrap operation (as knife delegates to ssh to do the actual work).