add actions to sensors in js gui, including links to copy the value, and to copy the REST endpoint. to use, simply hover over the value then go to the caret pointing down.
there is a lot of sad complexity here because copy is not possible in JS for security reasons;
it's only possible in flash, and then only when the flash embedded item has been clicked.
ZeroClipboard (developed by github) makes this as easy as it can, but when combined with
a popup menu there is still a mess, because the ancestor elements (eg the float group / popup menu)
lose their hover when the embed is hovered over.  so there is some horrible logic to keep the
menu up while the user navigates around.

i've also tweaked the mRender routine to minimise menu recomputes, so it prefers to update the html
of the cell rather than regenerate that html, meaning that the menu is not wiped while the user is
using it.

in short when you need to copy a hostname or something to the clipboard, this tries to make it easy!