Fix for brooklyn.rest.resources.SensorResourceTest.testGetPlain error.
Failing with expected "12345 frogs" but got "12345". The tearDown in EntityResourceTest
was cleaning the renderer registry in the middle of executing SensorResourceTests' methods.
Turns out TestNG interleaves the tests from separate classes in a single thread (the execution is sequential).
EntityResourceTest seems like copy&pasted from SensorResourceTest with the tearDown
call not needed, removed tearDown and unrelated comments.