minor tweaks to catalog item id setting
move catalog-item inferencing based on entity proximity to a helper method with a bit of logging
(did this mainly because i didn't see it in addChild/addEnricher/etc but then i realized you'd
done it in setParent/setEntity ... now it's in both which is superfluous but harmless)

only change of note is putting {get,set}CatalogItemId in the EntityTransientCopyInternal so that
on hot-standby rebind we are allowed to set these. (not sure if that was a problem or not.)

next will test this, but need to grab some osgi items to do that!