Add the eclipse java compiler as an option for the whole project
Using a user provided compiler could lead to unexpected results (i.e. bugs, changing behaviour) so as an alternative add the option to compile the project with the eclipse compiler. It's much like the approach so far of using the eclipse-groovy-compiler, but not set as the default.

To compile successfully I had to rename a package, matching (case insensitive) to an existing class. Looks like OS X's case insensitive file system is at the root (probably an issue under windows as well). https://jira.codehaus.org/browse/PLX-458.

To use add -Peclipse-compiler to the maven command.