Support scriptDir location config
This modifies SshMachineLocation to propagate certain config keys to
flags on SshTool execCommand and execScript invocations; the first use
of this is to propagate the "scriptDir" config. JcloudsLocation is also
modified to propagate this configuration. The result is that "scriptDir"
can be specified in the configuration for a jclouds-based location, and
the value of this property will be passed to every invocation of
SshTool.execScript().

The use case for this is to allow Brooklyn to be usable on locations
where the VM image mounts "/tmp" with "noexec".

Also adds some integration tests to verify the correct operation of the
scriptDir flag.