Adds EntityAndAttribute.create(...)
It’s nicer to use than the constructor because you don’t need to
specify generics and don’t get warned about it. However, not everyone
is used to such static methods so leaving the constructor as well
for ease of discovery.

Deprecates the old supplier() method, which should really have been 
called create().