Don't create install dir in copy-pre-install-resources unless there are resources to install.
Lots of unit tests assume that no ssh connection will be established when testing localhot entities. The addition of copy-pre-install-resources breaks the tests if local ssh is not configured properly. Create install dir only if there are resources to copy (and therefore the test expects it will try to ssh).