BrooklynYamlLocationResolver: don’t swallow exception
Previously inresolveLocation() we were just getting the “collapseText”
of the exception in the thrown toString, rather than having the
cause set. This meant that the underlying cause (e.g. of an NPE etc)
was never being logged.

Instead, set the cause when throwing the the IllegalStateException,
as well as having the nice toString.