Don't try to re-submit DST secondary tasks if already executed
Effectors invoked against an entity will be submitted twice - once in the entity's context and again in the TaskQueueingContext of the caller (see brooklyn.management.internal.LocalManagementContext.runAtEntity(Entity, TaskAdaptable<T>)). Usually it's not a problem because the ExecutionContext will notice it's already submitted and ignore it. But if the entity is unmanaged in the time being then the call to get the entity's ExecutionContext will fail.