Fix assertions for 'succeeds eventually' behaviour.
The following test case will not work with the code as currently defined:

```yaml
name: testcase1
location:
  mylocation
services:
- type: org.apache.brooklyn.entity.webapp.tomcat.TomcatServer
  id: tomcatServer
  timeout: 15m
- type: org.apache.brooklyn.test.framework.TestCase
  name: TomServTest
  brooklyn.children:
  - type: org.apache.brooklyn.test.framework.TestSensor
    target: $brooklyn:component("tomcatServer")
    sensor: service.isUp
    timeout: 1m
    assert:
    - equals: true
```

With the two entities (server and test) deployed independently instead of
within the same test case, the "checkAssertions" code will fail as it collects
all failures in AssertionSupport, when it should only collect those that
occur from the "Asserts.succeedsEventually".

Changing this code in this commit to allow tests to work.  The effect will
be that only one failed assertion will be reported on, even if many
assertions might fail. We can come back to this in order to see if multiple
assertion failures may be collected, but this is not necessary now.