[BROOKLYN-183] Remove karaf dependencies.xml from git
The dependencies.xml file can be used to track dependency changes for
karaf features. For this reason it has to be added to the git repository
so such changes can be detected between commits.

However, developers complained about having this file on git, since it
confusingly keeps showing up in commits. Also, at this stage this
tracking is unnecessary.