group software process tasks better
causes a much nicer tree to display, rather than lots of steps.
most are no-op and i think this structure makes it better to navigate.
unfortunately it is hard to tell if a step is no-op, as they are methods,
so we don't have a way just to not run empty steps.
it would be better to do away with most of them, finding a different mechanism
for subclasses to insert at the right point, but this helps in the short term.