BROOKLYN-231 fix race where updating and deleting can cause deleted item not to be deleted
the persists and deletions run simultaneously; now, don't write things which are deleted.
also suppress the immediate delta write that was previously being done;
this may impact tests if there relying on an immediate checkpoint, but that's a bad test,
for in normal operation checkpointing should be started when the mgmt node starts and there should be nothing to persist.
if we support turning on persistence later that should trigger a full write not a delta write.