Add ability to scan catalog.bom files in the root of bundles.
This adds entries to the catalog from any catalog.bom files in the bundle,
automatically adding a 'brooklyn.libraries' section for the bundle itself
to the bom before scanning it for items.

e.g.

brooklyn.catalog:
    version: 0.9.0
    scanJavaAnnotations: true
    description: For testing loading catalog.bom
    items:
    -   id: simpleTest1
        item:
            type: org.apache.brooklyn.test.osgi.entities.SimpleEntity

Note, however that the scanJavaAnnotations will not work on the bundle. Only
explicitly declared items will be added to the catalog.