softwareProcess.restart: even on error, set expected=running
Previously if the restart failed in some way (e.g. an error in the 
launch script or a timeout waiting for serviceUp), we would leave the 
expectedState as “starting”. This meant that serviceUp=false was not
rendered as an error.

Always setting expected=running is better, but is not great. For 
example, if a script fails in restart then we won’t mark the 
entity as onFire if the basic service-up checks still pass. This is
different from the start() behaviour. But then we are doing less than
in start (e.g. in start a script failure could mean that a WAR was not
installed so the simple serviceUp reachability check would not be
sufficient if customize() had failed).