Add catalog.bom application white/blacklists.
Any bundle with a catalog.bom file may specify applications to
appear in the catalog (itemType: template).

It is desirable to have a config time mechanism to control what
bundles are allowed to present applications in the catalog and
what aren't.

This change adds a white/blacklist mechanism to the catalog.bom
bundle scanning. Each list takes the form of a CSV whitelist of
regexes to match against bundle symbolic ids.  A bundle will be
allowed to add an application (template) to the catalog if its
bundle id matches one of the regexes on the whitelist, and also
does not match one of the regexes on the blacklist.

The configuration keys and default values for the lists are:

brooklyn.catalog.osgi.application.whitelist=.*
brooklyn.catalog.osgi.application.blacklist=

(i.e. by default accept all bundles).

these values are specified in the file org.apache.brooklyn.core.catalog.bomscanner.cfg

Note that two 'default' example applications are supplied in the bom
for brooklyn-software-base.