Add Karaf equivalents of cloud commands.
Refactors the CloudExplorer class into a core CloudExplorerSupport class
that does the work without knowing about any command line functionality,
plus a smaller CloudExplorer that does the command line handling and
invokes the CloudExplorerSupport to do the work.

Adds Karaf commands with the same names that use Karaf's command line
support to invoke the CloudExplorerSupport.

Adds the Karaf commands bundle as a boot feature so the commands are
readily available on startup.