Transformer enricher: support triggerSensors
Some uses of the transformer want to be triggered by multiple sensor
events (e.g. when they will combine two sensor values). This change
allows us to subscribe to multiple sensors, and to trigger the
evaulation when any of those change.

The sourceSensor is now optional. You can use it in the same way as
previously, or you can use just triggerSensors, or you can combine the
two. If combined, then the triggerSensors will cause us to look up the
latest value of the sourceSensor and synthesise an event as though
the source sensor had just published.