Test and fix locations not being unmanaged
Only MachineLocation locations created by SoftwareProcess are unmanaged. Currently there's no owner for LocationResolver locations so there's no one to unmanage them.

The changes will mark locations created as part of bringing up a spec as owner by the entity where they are attached. When the entity is unmanaged the locations will be unmanaged as well.