Ensure detection of failed pre.install.command and related commands.
Currently 'pre.install.command' and related steps do not fail if the command(s) returns a non-zero exit.
This can mean your install will fail but Brooklyn won't detect it (perhaps until some subsequent stage, or not at all).

This change adds detection of the return code and failure if it is non-zero. The 'contract' for the methods is added as a comment.  The AbstractSoftwareProcessSshDriver is changed to detect the failure and throw an exception. The AbstractSoftwareProcessWinRmDriver already does the right thing and doesn't need changed.

While this could break some existing blueprints, I think in such cases it's more likely that it is highlighting a problem that has been missed, rather than causing a problem because someone is explicitly relying on that behaviour.