Updates for effector calling.
this includes the following changes:

An "avoidSideEffects" capability in ValueResolver, to allow callers to
specify that any tasks which are marked as having side effects (a new
marker Interface) should be ignored.  The use case for this is illustrated
by its use here in  AbstractConfigurationSupportInternal in the
getNonBlocking method invoked by ConfigConstraints.validateAll.  This is
called during validation of the YAML, and evaluates all config key values,
which without this change would result in unwanted repeat executions of
effectors (and at a point when their target entities have not yet been started).
Similarly it is called during calculation of a hash for the extra salt for
VanillaSoftwareProcessSshDriver, again a situation where we do not want
effectors being called.

An additional effector() method on BrooklynDslCommon, to add support for
varArgs passing of the effector arguments, and a matching new method on
BrooklynDslDeferredSupplier.

The task created by ExecuteEffector is cached, in order to avoid unwanted
repeated executor invocations,  as illustrated in the test
testEffectorCalledOncePerConfigKey in EffectorsYamlTest.

test-app-with-effectors.yaml is removed as it's easier to see what's going
on with the YAML inline within the tests.

New tests are added in EffectorsYamlTest.

A new 'sequenceEffector' is added, just added directly  to TestEntity, in case
it happens to be useful for anyone else.