Fix etcd connection failure for different external and subnet networks.
Builds on https://github.com/cloudsoft/container-service/pull/55,
do not merge until after that.

Adds subnet address as as subjectAltName as well as host address.

This is because on Softlayer or other clouds where the public
address is on a different network from the subnet address,
swarm fails to connect to the etcd discovery mechanism.

The problem is that the public host address is configured in the
certificate but the etcd config is using the subnet address for the
client and peer urls. When swarm nodes connect on the subnet
address the certificate presented does not match the IP address
requested and the connection fails.

By adding the subnet address as another alternative name it
is possible to connect using either address.