[TODO] Set provisioning properties for Swarm nodes
Adds Swarm minRam and minCores configuration.

Currently does not work due to issues resolving the
provisioning.properties config keys in the JcloudsLocation
context, which does not appear to have an entity set.

If the entire propery map is use as a configuration
value, the coercion fails as it is only trying to
coerce the individual emtries in the map.