Fix loc.getHostname when DSL for credentials
If using DSL $brooklyn:config(“password”) for machine’s creds, then
if call machine.getHostname() outside of the entity’s context it will
fail to resolve the password so will throw an exception.

Instead, we now defer calling getLoginCredentials(), to only do it for
AWS (where we ask it for the hostname), and then if there is an 
exception we fall back to the defaults.