Test executing a task in a context different from its own
When executing a task which is already tagged with a context entity in the execution context of another task, the task ends up with two context entities in its tags. It's indeterministic which one will be available through BrooklynTaskTags.getContextEntity. Add tests to cover the expected behaviour (which are failing atm) - the task should always have a single context entity, the one explicitly applied on the task before submitting.