Stop persitence before rebind attempt
This is an attempt to fix test failures on the Apache Jenkins Windows slaves. The tests would frequrently fail with "The process cannot access the file because it is being used by another process" when trying to read the persisted state. Persistence in the old management context wasn't being stopped before trying the rebind so there's a chance that it would try to overwrite the files while the new management context rebinds, leading to file locking errors. Stopping persistence before rebind should solve the locking failures.