Fix app.stop so doesn’t go on-fire briefly
Previously in app.stop, we set the serviceUp=false before setting
expectedState=stopping. This meant there was a race where we could
briefly infer the actual state as on_fire.

This caused failure of SoftwareProcessEntityLatchTest.testStopLatchBlocks:
the attributeWhenReady latch would abort because it received an
event saying that the app was on_fire, so the stop_latch completely
immediately.

The fix is simple: in app.stop, set the expectedSate=stopping before
setting serviceUp=false.