Initialize REST API security correctly for integration tests
Also add checks for it in the tests.

Tests had different behaviour depending on whether brooklyn-ui existed along brooklyn-server on disk. When brooklyn-ui was there tests would find it and use the web app code from there, including the web.xml which restricts requests to only authorized users. When brooklyn-ui is missing though a default web app without web.xml was being created. This would result in non-authorized requests succeeding.  In this case requests including the authorization header would be accepted and rejected if the password is invalid. But in order to include the authorization header the server must first respond with a 401 which didn't happen. Moving web-security.xml to rest-server allows us to force request authentication for tests even if there's no web.xml in the web app.
Should be fine moving web-security.xml to rest-server (which is only used in classic) because it's only inserted by classic related code. Karaf Brooklyn doesn't have control over the web apps.