Fix out of order events triggered by ComputeServiceState
ComputeServiceState.onEvent should only be called in the event handler executor to keep the ordering of events.
Before the change what would happen is on calling setExpectedState(state) a "state.expected" event would be scheduled, with ComputeServiceState.onEvent(null) called in the same thread right after. This would create a race condition when calling setExpectedState in quick succession (onEvent called from the setExpectedState thread and from the event handler thread with different values).