LocalEntityManager: make fields private
Making these fields non-private is asking for trouble - it makes it 
harder to reason about what the class does and whether it’s safe, 
particularly around thread safety etc.

It is also exposing more implementation detail than is required, making
refactoring harder in the future.

If there is a compelling reason to provide access to the fields,
we can consider adding accessor(s).