solve problem with map eval where tasks are cancelled permanently
sets non-transient if a task is requested for a value-resolver

we might want to deprecate that altogether, instead use TaskFactory
so we can cancel things

don't think there will be much leaking because the ValueResolver isn't used for new tasks,
just for tasks which are set as values -- but we need to keep an eye on that.
such tasks should be cancelled when the entities are cleaned up.