Add findOpenSocketsOnNode to ReachableSocketFinder
This changes the semantics of ReachableSocketFinder. Previously it would
return as soon as any host+port was reached. Now it determines the
status of each socket in order. The checks are concurrent but, for
example, if the third of five sockets is reachable the status of the
first and second will also be determined.