transfer icons from catalog items to entities
and set icons on the common `server` and `cluster` items

previously while we do a lot to support icons and send detail via the API,
we don't _set_ them on entities unless explicitly in the definition.
since we often have them on catalog items, transferring them seems a sensible idea.

this adds two sample icons:

* server.svg from https://commons.wikimedia.org/wiki/File:Network-server.svg
             originally from the Tango project, released into the public domain

* cluster.svg from Font Awesome, SVG definition of the "server" icon, MIT license