allow reuse of embedded osgi framework (speed up tests and reduce memory)
primary change is in OsgiManager, to leave frameworks running

build times including test reduced from 17m to 11m on my box,
saving 6m in CAMP (from 7m to 1m)!

have done some assertions that starting a framework takes 200ms each time,
plus it leaks a couple MB, so it's a bad idea to bake this in to our unit tests.
(see comments in new OsgiTestingLeaksAndSpeedTest for full details.)

have also "confirmed" that reusing a running framework has no adverse impact,
after we remove bundles that a test installs (all tests pass in reusable mode
except a few which asserted the cache dir is deleted (as it no longer is)
and these still pass in non-reusable mode)