Revert "record VersionedName instead of string for containing bundle, as per PR"
This reverts commit f4205fde2c380a3193953c5305e398f974069111.

The "containingBundle" might, for a transitional period, need to support a reference to a catalog item instead of a bundle, in which case OSGi conventions might not be valid and we want this to be a string. TBC. Will add comment to that effect in the code shortly.

If/when reinstated we should also ensure the deserialization accepts strings (should probably also give a custom serializer so VersionedName is always written as a string to prevent bloat in the serialized format).