fix bundles set on scanning, and aborted work to scan osgi bundles
scanning for OSGi can be made to work in some cases if we have the jar,
but it won't work after rebind, or even in some cases if persisting (race deleting the JAR),
plus there are likely classloading issues;

note OSGi deliberately doesn't give us the JAR.

java scanning is discouraged by OSGi; it wants us to explicitly list things e.g. in the catalog.bom,
which is fine.