catalog.bom parsing can result in RegisteredType instances instead of CatalogItem
added to non-persisting BasicBrooklynTypeRegistry,
currently without validation;
with quite a few tests, and previous failing cobundle-reference tests now uncommented.
deletion support added as beta to TypeRegistry, works alongside Catalog deletion.

not done, lots of tests failing still, but lots working, and shape of code pretty good.
(even if the two-path code in BasicBrooklynCatalog is ugly, it's private and we'll be
able to simplify a lot if/when we cut the old catalog code)

need to next: perform some validation on the added registered types,
as a secondary phase so we can add all types first,
and populate supertypes then, which should allow all tests to pass,