remove FIRST_MEMBER sensor, it doesn't work, and deprecate FIRST
both are a bit flaky, but FIRST_MEMBER especially so:
it would never get cleared, and it could cause deadlock
(as parent attempts to update child's sensor while holding members lock,
meanwhile child might publish a sensor which causes parent subscriber to try to getMembers).

additionally we no longer set FIRST=(other entity) on the child, for the same reasons,
but also it causes conflict if the child is also a group (so FIRST might be the parent's FIRST
or its own FIRST).

the new "primary" model is a much better one to use instead when you need notification
of something being promoted, or an enricher if you need the parent's first injected locally
(but you should always be able to say `$brooklyn:parent().attributeWhenReady("first"))`)