fix type registry access synchronization and tweak minor load semantics
synchronization pointed out in PR, fixed.

also changed one bug, and it shares code with synch fixes so done here; the bug is:
until recently, if the same type were added from multiple bundles only the last one were kept.
on rebind the "last one added" might be different, so there was inconsistency.
previous commit means we keep all of them, and take the _first_ one added, and while
we could change it to have the same semantics (last one) it seems better to resolve
the rebind inconsistency, so now it takes from the "best" bundle, which is
the first alpha and preferring highest version.

use predictable resolution when multiple instances found