look at upgrade targets if type or bundle not found when rebinding instances and when deploying
this addresses most the places i found in the course of several tests.
there may well be others, and not all the places give good logging,
but those can be improved in the fullness of time
(and with yoml we have many fewer creation code paths so things get simpler!)

one thing that isn't addressed is within a persisted entity spec;
this unpacks the definition at deploy time and persists the unpacking,
rather than persisting the given entity spec yaml for use later.
i think this can be addressed later.