Update reflections dependency to 0.9.10
0.9.9-RC1 is unable to read classes containing lambda functions. This
meant that `brooklyn list-objects` did not output all classes. Refer to
https://github.com/ronmamo/reflections/issues/15 for futher context.

There is a more recent release of reflections that I think fixes further
issues (see GitHub) but it relies on Guava v20.
https://github.com/ronmamo/reflections/issues/194

Reflections' dependency on com.google.code.findbugs:annotations is
excluded because it clashes with a version from airline and should only
be needed at compile time anyway.