add an /applications/details endpoint which corrects problems with /fetch and adds new features
noticed that fetch actually recurses through all children entities which was never the intention and is wasteful at scale.
also for usability it would be nice to include tags and all/regex sensors and config values.  this adds support for that.

i've deprecated the existing `/fetch` call but left its behaviour unchanged.  have also expanded description of `/list` endpoint.

to make this easier i've promoted `getAll()` from SensorSupportInternal. previously we had no way on the public API to list all sensors for an entity.
it is still @Beta.