Allow apache-rat-pluin to work when building single modules
The configuration for apache-rat-plugin excludes the '**/target/**' path.
However, when building a single module, the basedir that the plugin uses by
default is the modules' target directory. The exclude signature for target
doesn't match, as it would be attempting to match a target directory under the
target directory itself. This works when building from a multi-module directory,
e.g. brooklyn or brooklyn-server, as the basedir in this case is not the target
directory.