Remove Config change listener
This was causing a memory leak as every request creates a new
DelegatingSecurityProvider and this registers a config change listener.
On config change the delegate security provider would then attempt a
reload 1000s of times although only one would actually have effect and
the rest would then use the cached delegate.

This wasn't needed and the functionality remains the same - minus the
leak of memory.