start feeds once entity is fully managed, not before
otherwise isManagedActive test is false if not fully managed.

also subscribe to service up if feed is enabled only when service is up, and misc tidies.

but might be better to leave this as was, and instead use isManagedActiveOrComingUp (and keep the other changes, just revert to starting feeds directly on feed creation. though this change seems better!)