support interpolation of a single dollar variable expression to resolve to an object
normally freemarker expects it to be outputtable to a writer;
however if writing eg `x = ${y}` we should allow y to resolve as an object.
(if writing `x = "hi ${y}"` then `y` needs to be outtable to a string.)

freemarker doesn't support this elegantly, so messy but small tweaks are needed:

* when model is accessed, we store the output in a thread-local map
* when output is needed, we check if the template is a single dollar variable and use reflection access

this mode of template processing needs to be specifically requested, and the
the thread local cache is explicitly set up and cleared for a processing cycle.