introduce and prefer `(type)` as the key to use in maps where `type` is ambiguous
previously we used `@type` as the key if the object had a field `type`.
this is still supported but so is `(type)` now with the latter preferred.
the reason for this is that `(type): Class` is allowed in yaml whereas
if a key starts `@` it is necessary to quote it, eg `"@type": Class`.

we also add support for `(type)` in several places, BeanWithType and CAMP and JavaClassName transformers,
so we can use it better in registered types.

introduce new test which failed before and passes with the changes.