better deserialization for adjunct specs
was picky about use of registered types vs classes, brooklyn.config vs flags;
now jackson deserialization handles most of those, understanding specs and registered spec types,
with lots of tests. other minor tweaks for use of specs eg in wrapped values.

also fixes "uniqueTag" in addPolicy step so that it bases on retention hash,
and retention hash takes a hash of steps if there is no name, so it is consistent