fixing osgi loading issue where osgi lets resources come from other bundles than oneself
if package imports include its own packages and there are eg bundle v 1-A and 1-B

always prefer local resources. (might not do that for classes, but the main issue was during bom processing.)

also be lenient if such a problem does occur, but there are no types affected, warn rather than throw.