let DslPredicate use wrapped values so we can inject supplied values into this
workflow uses this so that conditions can be parsed once and processed subsequently.
if too messy we might scratch that, always do late resolution.
primarily an issue for foreach and custom where the condition needs to be evaluated
after scratch variables are inserted.  (that could be handled another way which might be simpler.)

but the abilities introduced here, for conditions to be more dynamic are useful.

however they currently store the context which is not good.