store whether a bundle comes from the initial catalog, and don't persist those
change previous commit so that we don't exclude persisting of things from URL's,
but we do exclude persisting of things from the catalog by default.

we needed to persist things from URL's because we might manually install bundle A with libraries B,
and we need to persist a record that those libraries B are installed and persisted,
because if we remove A, we don't remove B, but if we haven't persisted a record of B, then B is removed after rebind.

we could record an XML record without storing the JAR, but that is more work;
and it makes more sense that things in the catalog we expect to be defined in the catalog;
if you want to keep those on a restart, they need to stay in the catalog (or have upgrade instructions in catalog),
or be installed manually prior to restarting, if you want to remove it from the catalog while it is in use,

installing manually causes things to be persisted even if they are already in the catalog.