Change config/sensor to return 200, 204, or 404. Tidy other API documentation bits.
            @ApiResponse(code = 200, message = "OK"),
            @ApiResponse(code = 204, message = "No Content. The config is known, but unset."),
            @ApiResponse(code = 404, message = "Could not find application, entity or config key"),

The behavior is now as above.  404 is returned if the sensor (or config) is not defined, and 204 if it is defined but unset.
200 with null or empty string is only returned if the sensor/config is explictly set as such.

Additionally, some discrepancies where Accepted was used instead of Created or sometimes OK has been fixed,
and OK is returned in some places where ACCEPTED or CREATED had been where it makes more sense to do so.